Formatted all the code

+++ Libs/LibC/musl/complex/cimagl.c
@@ -1,6 +1,5 @@
 #include "libm.h"
 
-long double (cimagl)(long double complex z)
-{
-	return cimagl(z);
+long double(cimagl)(long double complex z) {
+    return cimagl(z);
 }
