+++ Libs/LibC/musl/complex/creall.c
@@ -1,6 +1,5 @@
 #include <complex.h>
 
-long double (creall)(long double complex z)
-{
-	return creall(z);
+long double(creall)(long double complex z) {
+    return creall(z);
 }
